Start migrating users endpoints to new pattern
